requirements.txt
fastapi
uvicorn
requests
pandas
certifi
httpx
python-jose[cryptography]